+++ test/awscdk/python-app.test.ts
@@ -1,4 +1,4 @@
-import { AwsCdkPythonApp } from "../../src/awscdk";
+import { AwsCdkPythonApp, AwsCdkPythonAppOptions } from "../../src/awscdk";
 import { synthSnapshot } from "../util";
 
 test("create cdk python project", () => {
@@ -80,4 +80,27 @@ test("CDK v1 deps", () => {
     '# ~~ Generated by projen. To modify, edit .projenrc.py and run "npx projen".'
   );
   expect(snapshot["requirements-dev.txt"]).toContain("projen==99.99.99");
+});
+
+test("awscdk app is set up with custom appEntrypoint", () => {
+  const customAppEntrypoint = "custom_app.py";
+
+  const options: AwsCdkPythonAppOptions = {
+    authorName: "Test Author",
+    authorEmail: "test@example.com",
+    moduleName: "test_module",
+    name: "test_project",
+    version: "0.1.0",
+    cdkVersion: "1.0.0",
+    appEntrypoint: customAppEntrypoint,
+  };
+
+  // Instantiate AwsCdkPythonApp with custom options
+  const app = new AwsCdkPythonApp(options);
+
+  // Synthesize the project to generate the project structure
+  const snapshot = synthSnapshot(app);
+
+  // Check if the synthesized snapshot contains the custom app entrypoint
+  expect(Object.keys(snapshot)).toContain(customAppEntrypoint);
 });
